Login
To have something to discuss in interviews and what not, I'm plowing on with the login side of things in a seperate folder.

In terms of eventual implementation, I propose- to begin with even if it is a bit insecure- having a login screen overlaying the booking table in homeK5.html until a correct login is given.

Any thoughts other than the blatent security threats?